0.2.1 The biggest part of the logic in package /creation was relocated from Hero and Monster into AbstractCreation. Spring config was changed from xml to Java-Based

+++ src/main/java/abilities/AbstractAbility.java
@@ -38,7 +38,7 @@ public abstract class AbstractAbility implements IAbility {
         return damage;
     }
 
-    public int getRequiredLevel() {
+    public byte getRequiredLevel() {
         return requiredLevel;
     }
